+++ docs/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Changed
+
+* fix: retry logic now includes delays with exponential backoff. Retries should no longer happen too quickly for the replica to catch up.
+
+### Added
+
+* new `HttpAgent` option: `delayStrategy` - allows you to set a custom delay strategy for retries. The default is a newly exported `exponentialBackoff`, but you can pass your own function to customize the delay between retries.
+
 ## [1.2.0] - 2024-03-25
 
 ### Added